% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_variance.R
\name{Chr_region_annot}
\alias{Chr_region_annot}
\title{Chromosome physical segment annotation}
\usage{
Chr_region_annot(file_tmp, txdb_file)
}
\arguments{
\item{file_tmp}{Chromosome physical segment data, chrXX, xxx, xxx format}

\item{txdb_file}{Genome annotation databases, such as TxDb.Hsapiens.UCSC.hg19.knownGene}
}
\value{
Annotated chromosome physical segment data
}
\description{
Chromosome physical segment annotation
}
