% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{Cov_circle}
\alias{Cov_circle}
\title{According to HG19, the circle chart is displayed}
\usage{
Cov_circle(cov_file)
}
\arguments{
\item{cov_file}{cov format data generated by bismark}
}
\value{
Circle Chart
}
\description{
According to HG19, the circle chart is displayed
}
