% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{Cov_to_text}
\alias{Cov_to_text}
\title{Cov files to text files}
\usage{
Cov_to_text(file_dir, region_split_file)
}
\arguments{
\item{file_dir}{cov format data generated by bismark}

\item{region_split_file}{Processed physical fragment data, chrXX, xxx, xxx format}
}
\value{
DNA methylation or chromatin accessibility matrix
}
\description{
Cov files to text files
}
