% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_variance.R
\name{GSEA_analysis}
\alias{GSEA_analysis}
\title{GSEA analysis}
\usage{
GSEA_analysis(file_tmp, gmt_file, pvaluecutoff = 1)
}
\arguments{
\item{file_tmp}{Difference analysis results}

\item{pvaluecutoff}{adjusted pvalue cutoff}

\item{gmt_data}{user input annotation of TERM TO GENE mapping, a data.frame of 2 column with term and gene}
}
\value{
gseaResult object
}
\description{
GSEA analysis
}
