% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{Meth_colname_sum}
\alias{Meth_colname_sum}
\title{Statistics on the number of meths and UNmeths at different stages}
\usage{
Meth_colname_sum(file_tmp, string)
}
\arguments{
\item{string}{}
}
\description{
Statistics on the number of meths and UNmeths at different stages
}
