% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_variance.R
\name{Meth_group_variance_analysis}
\alias{Meth_group_variance_analysis}
\title{Group analysis}
\usage{
Meth_group_variance_analysis(
  file_data,
  file_group,
  group_id1,
  group_id2 = NULL,
  method = "t.test"
)
}
\arguments{
\item{file_group}{Group Information}

\item{group_id1}{Group Character1}

\item{group_id2}{Group Character2}

\item{file_tmp}{DNA methylation or chromatin accessibility level data}

\item{mothed}{Statistical methods}
}
\value{
Difference analysis results
}
\description{
Group analysis
}
