% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{Meth_level_stage_count}
\alias{Meth_level_stage_count}
\title{The number of methylation levels in the same stage was counted through the .cov file}
\usage{
Meth_level_stage_count(dir, level_string)
}
\arguments{
\item{dir}{cov format data generated by bismark}

\item{level_string}{Different sample information or developmental stages}
}
\value{
Data information of the same sample group or the same developmental stage
}
\description{
The number of methylation levels in the same stage was counted through the .cov file
}
