% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation.R
\name{MOFA_est}
\alias{MOFA_est}
\title{MOFA object is created}
\usage{
MOFA_est(file_tmp, group = NULL)
}
\arguments{
\item{file_tmp}{DNA methylation or chromatin accessibility level data}

\item{group}{Sample grouping data}
}
\value{
a trained MOFA object
}
\description{
MOFA object is created
}
