% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{RNA_enrichment_analysis}
\alias{RNA_enrichment_analysis}
\title{Enrichment analysis of RNA expression matrices}
\usage{
RNA_enrichment_analysis(file_tmp, method = "GO")
}
\arguments{
\item{file_tmp}{Expression matrix derived from differential analysis}

\item{method}{Enrichment methods}
}
\value{
Enrichment results
}
\description{
Enrichment analysis of RNA expression matrices
}
