% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA.R
\name{RNA_group_variance_analysis}
\alias{RNA_group_variance_analysis}
\title{Differential analysis of RNA expression matrices}
\usage{
RNA_group_variance_analysis(
  file_tmp,
  file_group,
  group_id1,
  group_id2 = NULL,
  method_model = "wilcoxon"
)
}
\arguments{
\item{file_tmp}{RNA expression matrix}

\item{file_group}{Group Information}

\item{group_id1}{Group Character1}

\item{group_id2}{Group Character2}

\item{method_model}{Statistical methods}
}
\value{
Difference analysis results
}
\description{
Differential analysis of RNA expression matrices
}
